+++ EchoDesertTrips.Business.Entities/Optional.cs
@@ -17,10 +17,10 @@ namespace EchoDesertTrips.Business.Entities
         public string OptionalDescription { get; set; }
 
         [DataMember]
-        public float PricePerPerson { get; set; } 
+        public decimal PricePerPerson { get; set; } 
 
         [DataMember]
-        public float PriceInclusive { get; set; }
+        public decimal PriceInclusive { get; set; }
 
         [DataMember]
         public bool Visible { get; set; }
