+++ EchoDesertTrips.Business.Entities/Reservation.cs
@@ -30,9 +30,9 @@ namespace EchoDesertTrips.Business.Entities
         [DataMember]
         public int? AgentId { get; set; }
         [DataMember]
-        public double AdvancePayment { get; set; }
+        public decimal AdvancePayment { get; set; }
         [DataMember]
-        public double TotalPrice { get; set; }
+        public decimal TotalPrice { get; set; }
         [DataMember]
         public DateTime PickUpTime { get; set; }
         [DataMember]
